pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slack-chat-migrator"
description = "Tool for migrating Slack exports to Google Chat"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Nick Lamont", email = "nick@nicklamont.com" }]
keywords = ["slack", "google-chat", "migration", "workspace", "chat", "data-migration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Utilities",
]
dynamic = ["version"]
dependencies = [
    "emoji~=2.15",
    "requests>=2.32,<3.0",
    "PyYAML>=6.0,<7.0",
    "google-auth>=2.40,<3.0",
    "google-api-python-client>=2.177,<3.0",
    "google-auth-httplib2~=0.3",
    "google-auth-oauthlib>=1.2,<2.0",
    "click>=8.1,<9.0",
    "rich>=13.0,<14.0",
]

[project.urls]
Homepage = "https://github.com/nicklamont/slack-chat-migrator"
"Bug Tracker" = "https://github.com/nicklamont/slack-chat-migrator/issues"
"Source Code" = "https://github.com/nicklamont/slack-chat-migrator"

[project.scripts]
slack-chat-migrator = "slack_chat_migrator.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "mypy>=1.10",
    "types-requests>=2.32",
    "types-httplib2>=0.22",
    "types-PyYAML>=6.0",
    "pytest-cov>=5.0",
    "coverage>=7.0",
    "pre-commit>=3.0",
    "commitizen>=3.0",
]
setup = [
    # All setup operations use the REST discovery API (google-api-python-client),
    # which is already a core dependency. This extra is kept for forward compatibility.
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]

[tool.setuptools.dynamic]
version = { attr = "slack_chat_migrator.__version__" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["slack_chat_migrator*"]

[tool.setuptools.package-data]
slack_chat_migrator = ["py.typed"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF", "C90", "S"]
ignore = ["E203", "E501", "S101", "S110", "S112"]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S106", "S108"]

[tool.ruff.lint.isort]
known-first-party = ["slack_chat_migrator"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = ["emoji.*", "slackblocks.*", "googleapiclient.*"]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests"]
# B101: assert used for type narrowing
# B110/B112: intentional error suppression in migration recovery logic
# B324: MD5 used for file deduplication, not security
# B404/B603: subprocess used for gcloud bootstrap and clipboard (hardcoded commands)
skips = ["B101", "B110", "B112", "B324", "B404", "B603"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: Integration tests (dry-run, no credentials needed)",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version = "0.3.2"
version_files = ["src/slack_chat_migrator/__init__.py:__version__"]
update_changelog_on_bump = true

[tool.coverage.run]
branch = true
source = ["slack_chat_migrator"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 90
# CLI orchestration modules are omitted because they wire together
# already-tested services and require full end-to-end setup (credentials,
# Slack export, Google API access) that unit tests cannot provide.
omit = [
    "src/slack_chat_migrator/cli/migrate_cmd.py",
    "src/slack_chat_migrator/cli/cleanup_cmd.py",
    "src/slack_chat_migrator/cli/setup_cmd.py",
    "src/slack_chat_migrator/services/setup/*.py",
]
